Move more classes
