+++ include/RAJA/IndexSetBuilders.hxx
@@ -55,7 +55,7 @@
 
 #include "RAJA/config.hxx"
 
-#include "RAJA/int_datatypes.hxx"
+#include "RAJA/Types.hxx"
 
 namespace RAJA
 {
