+++ include/RAJA/internal/BaseSegment.hxx
@@ -55,7 +55,7 @@
 
 #include "RAJA/config.hxx"
 
-#include "RAJA/internal/int_datatypes.hxx"
+#include "RAJA/Types.hxx"
 
 namespace RAJA
 {
