+++ include/RAJA/internal/IndexSetUtils.hxx
@@ -56,7 +56,7 @@
 
 #include "RAJA/config.hxx"
 
-#include "RAJA/internal/int_datatypes.hxx"
+#include "RAJA/Types.hxx"
 
 namespace RAJA
 {
