+++ include/RAJA/internal/PolicyBase.hxx
@@ -1,9 +1,6 @@
 #ifndef RAJA_POLICYBASE_HXX
 #define RAJA_POLICYBASE_HXX
 
-#include "RAJA/internal/int_datatypes.hxx"
-#include "RAJA/internal/RangeSegment.hxx"
-
 namespace RAJA {
 
 struct PolicyBase { };
