+++ include/RAJA/internal/exec-sequential/forall_sequential.hxx
@@ -58,7 +58,7 @@
 
 #include "RAJA/config.hxx"
 
-#include "RAJA/internal/int_datatypes.hxx"
+#include "RAJA/Types.hxx"
 
 #include "RAJA/internal/fault_tolerance.hxx"
